fixed a bug in estimateProportion.default and in plotCorrelationDifference
estimateProportion.default - when na.rm was false, NAs were being counted not only in the total N but also in each case count.  Fixed.

plotCorrelationDifference - had fixed ylims of -1.1 and 1.1, and this was removing difference CI when it was out of bounds.  added a calculation to set ylims based on max of 1.1 and difference CI and min of -1.1 and difference CI.