Added datasets
If you don't add datasets in this way, they won't show up in the Data Library.
At a later point, you may be interested in writing the datafiles out from R directly. I wrote a R-package for that (jmvReadWrite on CRAN). The package also includes a function `describe_omv` that permits giving the dataset a description. Let me know, if you want to to that and need help.